PHP Web Assessment
Here is my version of the Oxford Club PHP Web Assessment. It includes 3
pages in the navigation Home, Product, and About. When a user submits
there email on the home page they are taken to the thank you page of
which ever newsletter they signed up for. If they signed up for both
the first newsletter is registered then the second.